fix warnning & update gml.

+++ got2d/source/render_system.h
@@ -140,14 +140,14 @@ public:
 	virtual void SetVSConstant(unsigned int index, float* data, unsigned int size, unsigned int count) override;
 	virtual void SetPSConstant(unsigned int index, float* data, unsigned int size, unsigned int count) override;
 	virtual g2d::Texture* GetTexture(unsigned int index) const override { return m_textures[index]; }
-	inline virtual unsigned int GetTextureCount() const override { return m_textures.size(); }
+	inline virtual unsigned int GetTextureCount() const override { return static_cast<unsigned int>(m_textures.size()); }
 	inline virtual const float* GetVSConstant() const override { return reinterpret_cast<const float*>(&(m_vsConstants[0])); }
-	inline virtual unsigned int GetVSConstantLength() const override { return m_vsConstants.size() * 4; }
+	inline virtual unsigned int GetVSConstantLength() const override { return static_cast<unsigned int>(m_vsConstants.size()) * 4; }
 	inline virtual const float* GetPSConstant() const override { return reinterpret_cast<const float*>(&(m_psConstants[0])); }
-	inline virtual unsigned int GetPSConstantLength() const override { return m_psConstants.size() * 4; }
+	inline virtual unsigned int GetPSConstantLength() const override { return static_cast<unsigned int>(m_psConstants.size()) * 4; }
 	inline virtual void Release() override { delete this; }
 	inline unsigned int GetTextureCount() { return static_cast<unsigned int>(m_textures.size()); }
-
+	
 private:
 	std::string m_effectName;
 	std::vector<g2d::Texture*> m_textures;
